+++ client/src/components/RegistrationForm/registrationFrom.scss
@@ -26,4 +26,9 @@
       margin-top: 10px;
     }
   }
+
+  &--error {
+    color: red;
+    text-align: center;
+  }
 }
