+++ client/src/pages/registration/RegistrationPage.jsx
@@ -1,6 +1,8 @@
 import React from "react";
 import { useSelector } from "react-redux";
 import RegistrationForm from "../../components/RegistrationForm/RegistrationForm";
+import { main_route } from "../../shared/utils/_constans";
+import { NavLink } from 'react-router-dom';
 
 import './registration.scss';
 
@@ -8,8 +10,9 @@ export default function RegistrationPage() {
   const { isAuth, user } = useSelector((state) => state.login);
 
   if (isAuth) return(
-    <div>
+    <div className="LoginPage--container">
       <h1>{isAuth ? `Welcome, ${user.email}!` : 'Log in, please'}</h1>
+      <NavLink to={main_route}><button className='btn btn--orange'>Let's Shopping!</button></NavLink>
     </div>
   );
