+++ database/users/tokens.json
@@ -1,1 +1,1 @@
-[{"id":"e3622bcf-4c45-41f3-94c1-1095eb9011ca","user":"862db23e-73f8-4fdb-9d24-fa8f83ff9e21","refreshToken":"eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJlbWFpbCI6InVzZXI0QGdtYWlsLmNvbSIsImlkIjoiODYyZGIyM2UtNzNmOC00ZmRiLTlkMjQtZmE4ZjgzZmY5ZTIxIiwibG9naW4iOiJsb2dpbjQiLCJhZG1pbiI6ZmFsc2UsImlhdCI6MTY1MTU4Njk1MSwiZXhwIjoxNjUyNDUwOTUxfQ.gqV0R_9kkB90e2rymasgijsPLzlW8Ps2m0QXxFX0ayg"}]
+[{"id":"a49b2d21-0c73-4a4e-9541-b88646f35ead","user":"862db23e-73f8-4fdb-9d24-fa8f83ff9e21","refreshToken":"eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJlbWFpbCI6InVzZXI0QGdtYWlsLmNvbSIsImlkIjoiODYyZGIyM2UtNzNmOC00ZmRiLTlkMjQtZmE4ZjgzZmY5ZTIxIiwibG9naW4iOiJsb2dpbjQiLCJhZG1pbiI6ZmFsc2UsImlhdCI6MTY1MTY1NzI2OSwiZXhwIjoxNjUyNTIxMjY5fQ.ztq5Ye_-DmX4qUcyviIIL5gQFDI2GjBvSKm3Uhxk2hg"}]
