+++ database/users/users.json
@@ -1,1 +1,1 @@
-[{"id":1,"login":"login","name":"name","phone":"+38 (222) 222 22 22","email":"email@qw.we","password":"qwe","role":"User","soc":[{"id":1,"name":"Facebook","link":"sasdasd"},{"id":3,"name":"Instagram","link":"instagram"}],"fav":[{"id":1}],"cart":[],"history":[{"id":1,"products":[]}]},{"id":2,"login":"shop","name":"shop","phone":"+38 (222) 222 22 22","email":"shop@qw.we","password":"qwe","role":"Shop","soc":[{"id":1,"name":"Facebook","link":"sasdasd"},{"id":3,"name":"Instagram","link":"instagram"}],"address":"address1"},{"id":"bc18fd3d-56ba-4d87-bb84-760a04b2dcf6","email":"user1@gmail.com","password":"$2b$04$3D3T4LEJ681cgLDk7K0tW.0eoz8VNBFWCOsrPg9Q5EVnMpQWUVD8u","admin":true,"cart":[{"id":"7fb8d374-4ee5-4579-af36-1967b60046a7","count":3}]},{"id":"73a3b824-6eb2-4aa4-956a-a7e71abda0c9","email":"user9@gmail.com","password":"$2b$04$SruaQMMBfZkrZ9n0QdziT.3JU5ZvGqyVPt0bf8Iy8wS0510vDBIme","admin":true,"cart":[{"id":"bf69b9b9-bfeb-4821-8623-cb174f8ff31d","count":2},{"id":"3db18d80-555e-4844-b0c4-3ec1ceb05a2e","count":1}]},{"id":"52201246-0818-4c74-9f63-44247e884a71","email":"user3@gmail.com","password":"$2b$04$ivxFYGIqc4boPsxLtvt5SeG6ERu9QGgpaRiFbrogKR3TMA2IYw7hG","admin":false,"login":"login3@gmail.com"},{"id":"862db23e-73f8-4fdb-9d24-fa8f83ff9e21","email":"user4@gmail.com","password":"$2b$04$clJ1rM.hDjBnBD2F7yTnFuRlbOi7Q0KQZJU8tgM.tI2hBqdeWHlQK","admin":false,"login":"login4"}]
+[{"id":1,"login":"login","name":"name","phone":"+38 (222) 222 22 22","email":"email@qw.we","password":"qwe","role":"User","soc":[{"id":1,"name":"Facebook","link":"sasdasd"},{"id":3,"name":"Instagram","link":"instagram"}],"fav":[{"id":1}],"cart":[],"history":[{"id":1,"products":[]}]},{"id":2,"login":"shop","name":"shop","phone":"+38 (222) 222 22 22","email":"shop@qw.we","password":"qwe","role":"Shop","soc":[{"id":1,"name":"Facebook","link":"sasdasd"},{"id":3,"name":"Instagram","link":"instagram"}],"address":"address1"},{"id":"bc18fd3d-56ba-4d87-bb84-760a04b2dcf6","email":"user1@gmail.com","password":"$2b$04$3D3T4LEJ681cgLDk7K0tW.0eoz8VNBFWCOsrPg9Q5EVnMpQWUVD8u","admin":true,"cart":[{"id":"7fb8d374-4ee5-4579-af36-1967b60046a7","count":3}]},{"id":"73a3b824-6eb2-4aa4-956a-a7e71abda0c9","email":"user9@gmail.com","password":"$2b$04$SruaQMMBfZkrZ9n0QdziT.3JU5ZvGqyVPt0bf8Iy8wS0510vDBIme","admin":true,"cart":[{"id":"bf69b9b9-bfeb-4821-8623-cb174f8ff31d","count":2},{"id":"3db18d80-555e-4844-b0c4-3ec1ceb05a2e","count":1}]},{"id":"52201246-0818-4c74-9f63-44247e884a71","email":"user3@gmail.com","password":"$2b$04$ivxFYGIqc4boPsxLtvt5SeG6ERu9QGgpaRiFbrogKR3TMA2IYw7hG","admin":false,"login":"login3@gmail.com"},{"id":"862db23e-73f8-4fdb-9d24-fa8f83ff9e21","email":"user4@gmail.com","password":"$2b$04$clJ1rM.hDjBnBD2F7yTnFuRlbOi7Q0KQZJU8tgM.tI2hBqdeWHlQK","admin":false,"login":"login4"},{"id":"cfc70b92-4f87-4311-82ae-70a7da32349a","email":"user5@gmail.com","password":"$2b$04$z7d/.9jW2IyjCfUgvJeM5.eGIXpi.wuK6BOlsDLFG3fygYlRgRy2O","admin":false,"login":"login5"},{"id":"6f71d851-5bb2-4052-9b95-64978ab8655e","email":"user7@gmail.com","password":"$2b$04$nqTmQXtu803gTl0La5dZGOZZvvpdPjZfBI/U10HSNzcJ1CpjxaIL.","admin":false,"login":"login7"},{"id":"3b6b2525-cee0-4bb5-b8f0-e1198f64a1ee","email":"user@gmail.com","password":"$2b$04$TYOZPokjxIciH1tOJPVpx.oZYKgdVWCBcKajBDiWK19CFQVI.7.Si","admin":true,"login":"login"},{"id":"5a46efe0-3a88-46c9-801e-d796ff043813","email":"user8@gmail.com","password":"$2b$04$qJJl3PTT/oau6f97jruV/.stNLBBF6I.urCgSyscNBAvNJCcgpndK","admin":false,"login":""}]
